Safari-611.1.10/LayoutTests/inspector/model/remote-object/number-expected.txt


-----------------------------------------------------
EXPRESSION: 0
{
  "_type": "number",
  "_description": "0",
  "_value": 0
}

-----------------------------------------------------
EXPRESSION: -0
{
  "_type": "number",
  "_description": "-0",
  "_value": 0
}

-----------------------------------------------------
EXPRESSION: 1
{
  "_type": "number",
  "_description": "1",
  "_value": 1
}

-----------------------------------------------------
EXPRESSION: -1
{
  "_type": "number",
  "_description": "-1",
  "_value": -1
}

-----------------------------------------------------
EXPRESSION: 1.234
{
  "_type": "number",
  "_description": "1.234",
  "_value": 1.234
}

-----------------------------------------------------
EXPRESSION: -1.234
{
  "_type": "number",
  "_description": "-1.234",
  "_value": -1.234
}

-----------------------------------------------------
EXPRESSION: 1e3
{
  "_type": "number",
  "_description": "1000",
  "_value": 1000
}

-----------------------------------------------------
EXPRESSION: Number.MAX_VALUE
{
  "_type": "number",
  "_description": "1.7976931348623157e+308",
  "_value": 1.7976931348623157e+308
}

-----------------------------------------------------
EXPRESSION: Number.MIN_VALUE
{
  "_type": "number",
  "_description": "5e-324",
  "_value": 5e-324
}

-----------------------------------------------------
EXPRESSION: NaN
{
  "_type": "number",
  "_description": "NaN",
  "_value": null
}

-----------------------------------------------------
EXPRESSION: Infinity
{
  "_type": "number",
  "_description": "Infinity",
  "_value": null
}

-----------------------------------------------------
EXPRESSION: -Infinity
{
  "_type": "number",
  "_description": "-Infinity",
  "_value": null
}

-----------------------------------------------------
EXPRESSION: 123n
{
  "_type": "bigint",
  "_description": "123n",
  "_value": "<filtered 123n>"
}
